+++ compiler/semmacrosanity.nim
@@ -146,7 +146,12 @@ proc annotateType*(n: PNode, t: PType; conf: ConfigRef) =
   of nkCurly:
     if x.kind in {tySet}:
       n.typ() = t
-      for m in n: annotateType(m, x.elemType, conf)
+      for m in n:
+        if m.kind == nkRange:
+          annotateType(m[0], x.elemType, conf)
+          annotateType(m[1], x.elemType, conf)
+        else:
+          annotateType(m, x.elemType, conf)
     else:
       globalError(conf, n.info, "{} must have the set type")
   of nkFloatLit..nkFloat128Lit: